[GearSwap] Flow Refactor Part 1 (trigger -> midcast)
GearSwap is suffering from some blurred lines. This commit merely seeks to
simplify the existing flow by defining some appropriate boundaries to alleviate
code duplication and minimize diversion from the main code path. It is crucial
that zero GearSwap behavior is modified to avoid impacting any existing user
files which may depend on even improper behavior.

GearSwap could benefit from standardizing its internal naming, such as using
'action' instead of ability, spell, etc. but since developers may be accustomed
to the current naming conventions, overhauling the naming scheme will be saved
for another time so reviewing this commit's changes is not too disorienting.
'Spell' is most commonly used in current GearSwap, so all player actions shall
be referred to as spells henceforth.

The roles of the various GearSwap components should be as follows:
outgoing text - Same as for any GearSwap trigger; use trigger-specific methods
to parse the source trigger and map it to valid spell and target tables.

flow_entry - Newly created to take over the role of building the spell table
and taking care of any trigger-agnostic details.

filter_pretarget - Gatekeeper for entering pretarget. Ensures GearSwap's
acquired info regarding the spell will result in a legal outgoing action packet
i.e. the player meets all requirements for using the requested spell.

filtered_action - Gives user function an opportunity to take action if the
requested spell fails filter_pretarget's checks.

pretarget - Should do as little as possible to minimize differences between the
'trigger -> pretarget -> precast' and 'trigger -> precast' code paths. Thus
pretarget should only be concerned with target related matters, mainly
finalizing the target after the user function an opportunity to change the
spell's target. Swapping gear in pretarget shouldn't even be allowed to ensure
the player gets the same sequence of gear swaps regardless of how they cast a
spell, but GearSwap is stuck with this behavior now.

filter_precast - Gatekeeper for entering precast. Ensures GearSwap's acquired
info regarding the target will result in a legal outgoing action packet.

precast - final preparations to fire the spell, including action packet
assembly and command registry invokation.

command_registry - track the spell throughout its life cycle.

Some comments:
Note that filter_pretarget and filter_precast in tandem perform exactly the
necessary checks to ensure GearSwap does not send an outgoing packet that the
client would not send (except for items, which suck). It should be the
responsibility of filter_pretarget to check if a requested item is in the
appropriate bag and/or is useable as this is directly analogous to checking if
the requested spell is in the player's spell menu and is useable. Unfortunately
it is necessary to know information about the target in order to know which
item checks need to be performed, but by definition the target is not finalized
until after pretarget, and filter_pretarget executes before pretarget. Thus
items suck and this commit will default to letting the outgoing item packet
assemblers perform the item checks that should fall into filter_pretarget.

The command registry exists to track the spell throughout its life cycle, but a
spell is not a spell until it is cast. Previously the command registry was
being invoked too early, as evidenced by an entry being created, copying global
values into it, then copying them back into global, then just deleting that
entry before doing anything with it in some cases. This commit only invokes
the command registry when it is truly needed; when a spell will be fired.

Similarly, GearSwap had a bit of a problem with premature packetulation where
packets would be assembled but then discarded in some cases. Idk maybe GearSwap
was HEXually frustrated.. they're never letting me submit another PR are they.

Monster skills are incompatible with the validabils table so they were removed.
validabils relies on the assumption that every spell name maps to a unique
index in its associated resource table, but this assumption is not true for
monster skills.

The function scheduler is used in place of pretarget_delayed_cast and
precast_delayed_cast to avoid creating another code path.

This commit fixes the following issues:
- /ma "Dia" 12345 resulted in an arithmetic error in packet assembly since
  filter_precast was not being called upon to do its job in all cases.
- Doing a /item trade when the target NPC is too far away will now properly let
  the command flow to outgoing text for the client to print the 'Target out of
  range' message instead of GearSwap faking it.
- If /ma "Dia" <t> was entered, and pretarget changed the target to <st>,
  GearSwap would output the original <t> to outgoing text instead of <st>
  resulting in a spell cast being initiated (while also skipping precast gear
  swaps) instead of being given an opportunity to select a subtarget. The
  st_flag would also be left set, so the next entered command would not receive
  any gear swaps.
- player.species will no longer contain invalid stale data if exiting
  monstrosity with PUP as main or sub job.
- maybe others